вынос логики effect в отдельный модуль
